+++ packages/yourloops/components/icons/diabeloop/right-icon.tsx
@@ -25,31 +25,29 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import * as auth0Mock from '@auth0/auth0-react'
-import { getAccessTokenWithPopupMock, logoutMock } from '../../mock/auth0.hook.mock'
-import { renderPage } from '../../utils/render'
-import { waitFor } from '@testing-library/react'
-import { AUTH0_ERROR_EMAIL_NOT_VERIFIED } from '../../../../lib/auth/models/auth0-error.model'
-import { AppRoute } from '../../../../models/enums/routes.enum'
-import { testVerifyEmail } from '../../use-cases/email-verification'
+import SvgIcon, { SvgIconProps } from '@mui/material/SvgIcon'
+import React, { ForwardedRef } from 'react'
+import { makeStyles } from 'tss-react/mui'
 
-describe('Verify email page', () => {
-  it('should display a description of the email verification process with options', async () => {
-    (auth0Mock.useAuth0 as jest.Mock).mockReturnValue({
-      isAuthenticated: false,
-      isLoading: false,
-      user: null,
-      getAccessTokenWithPopup: getAccessTokenWithPopupMock,
-      logout: logoutMock,
-      getAccessTokenSilently: jest.fn().mockRejectedValue({ error_description: AUTH0_ERROR_EMAIL_NOT_VERIFIED })
-    })
-    window.open = jest.fn()
+const styles = makeStyles()(() => ({
+  'checkmark': {
+    strokeWidth: 6,
+    strokeLinecap: 'round'
+  }
+}))
 
-    const router = renderPage(AppRoute.VerifyEmail)
-    await waitFor(() => {
-      expect(router.state.location.pathname).toEqual(AppRoute.VerifyEmail)
-    })
-
-    await testVerifyEmail()
-  })
+export const RightIcon = React.forwardRef((props: SvgIconProps, ref: ForwardedRef<SVGSVGElement>) => {
+  const { classes } = styles()
+  return (
+    <SvgIcon {...props} ref={ref}>
+      <svg viewBox="0 0 180 180" fill="none">
+        <rect width="180" height="180" rx="90" fill="currentColor"/>
+        <rect x="15.5" y="15.5" width="149" height="149" rx="74.5" fill="currentColor"/>
+        <rect x="15.5" y="15.5" width="149" height="149" rx="74.5" stroke="white"/>
+        <path d="M53.9465 98.6502L83.567 126.643L130.969 62.0462" stroke="white" className={classes['checkmark']}/>
+      </svg>
+    </SvgIcon>
+  )
 })
+
+RightIcon.displayName = 'RightIcon'