+++ CHANGELOG.md
@@ -5,6 +5,7 @@ It is based on Tidepool Blip 1.27.
 ## 3.0.6-beta - 2023-11-10
 ### Changed
 - YLP-2574 Create tool tips independently of the name of the columns
+- YLP-2749 Use a branded Verify Email result page instead of the Auth0 one
 
 ### Engineering use
 - YLP-2476 Use tide-whisperer v2: route /all.